fix(point-bounding-box): Fix distance calculation issues
Several issues were found in the code for the creation/calculation of point-bounding boxes. The
primary one was that we were reading from an incorrect variable, `@units` when deciding whether to
convert the input distance or not. There were a few other small bugs surrounding great-circle
distance calculations, as well as some cleanup. This change fixes all of these.